# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For Ride-With-Me",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/reservations/create": {
            "post": {
                "tags": [
                    "Reservation"
                ],
                "summary": "Create reservation",
                "description": "Create a new reservation",
                "operationId": "cb5c986faf5b25bd6a87358cd902758b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "route_id",
                                    "status",
                                    "seat"
                                ],
                                "properties": {
                                    "user_id": {
                                        "description": "ID of the user",
                                        "type": "integer"
                                    },
                                    "route_id": {
                                        "description": "ID of the route",
                                        "type": "integer"
                                    },
                                    "status": {
                                        "description": "Status of the reservation",
                                        "type": "string"
                                    },
                                    "seat": {
                                        "description": "Seat number",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Reservation created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Reservation"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Reservation already exists"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "An error occurred"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/reservations/update/{reservation}": {
            "put": {
                "tags": [
                    "Reservation"
                ],
                "summary": "Update reservation",
                "description": "Update the status of a reservation",
                "operationId": "d0483df571dd11a930c17cffd9ab7e27",
                "parameters": [
                    {
                        "name": "reservation",
                        "in": "path",
                        "description": "ID of the reservation",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "description": "Status of the reservation",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Reservation status updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Reservation"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Reservation not found"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "An error occurred"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/reservations/received": {
            "get": {
                "tags": [
                    "Reservation"
                ],
                "summary": "Get received requests",
                "description": "Get a list of received reservation requests",
                "operationId": "bdb6e848d2f8ded36eab288a54d7c005",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Reservation"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "An error occurred"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/reservations/sent": {
            "get": {
                "tags": [
                    "Reservation"
                ],
                "summary": "Get sent requests",
                "description": "Get a list of sent reservation requests",
                "operationId": "50305c2936b2e8fca39d97c930fa15b3",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Reservation"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "An error occurred"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Reservation": {
                "title": "Reservation",
                "description": "Reservation model",
                "required": [
                    "id",
                    "user_id",
                    "route_id",
                    "status",
                    "seat"
                ],
                "properties": {
                    "id": {
                        "description": "Reservation ID",
                        "type": "string",
                        "format": "uuid"
                    },
                    "user_id": {
                        "description": "ID of the user making the reservation",
                        "type": "string",
                        "format": "uuid"
                    },
                    "route_id": {
                        "description": "ID of the route being reserved",
                        "type": "string",
                        "format": "uuid"
                    },
                    "status": {
                        "description": "Status of the reservation",
                        "type": "string"
                    },
                    "seat": {
                        "description": "Seat number",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}